feat(core): move DeepestErrorFinder into dedicated deepest_error.rs
Issue #393 asks for DeepestErrorFinder to live in its own file
(crates/core/src/decode/deepest_error.rs) rather than inside
chain_analyzer.rs where it was implemented for #378. Extracts the
struct, finder, and tests into the dedicated module, re-exporting via
decode::mod, and fixes an off-by-one in the reported call depth: the
failure depth was read from the post-push stack length instead of the
depth of the frame that was actually active when the failure fired,
which under-reported the true depth by one on every nested (2+ level)
failure. Adds coverage for 4-level nesting and recursive re-entry.
