+++ crates/core/src/decode/mod.rs
@@ -7,6 +7,7 @@ pub mod contract_error;
 pub mod contract_error_resolver;
 pub mod cross_contract;
 pub mod decode_context;
+pub mod deepest_error;
 pub mod diagnostic;
 pub mod event_walker;
 pub mod fee_analyzer;
@@ -27,6 +28,7 @@ pub use auth::{
 };
 pub use auth_address_nonce::AddressWithNonce;
 pub use chain_analyzer::{analyze_call_chain, CallChain, ChainAnalyzer, ChainFrame, FrameRole};
+pub use deepest_error::{find_deepest_error, DeepestError, DeepestErrorFinder};
 pub use function_call_decoder::{DecodedArgument, DecodedFunctionCall, FunctionCallDecoder};
 pub use multi_op_decoder::{decode_transaction_with_op_filter, MultiOpDecoder};
 pub use resource_analyzer::{
